WIP: add concrete x86 semantics 'IsValue' type.
https://galois.atlassian.net/browse/RAD-29

This mostly a copy of the existing 'IsValue' 'Expr' type in
'Reopt.Semantics.Implementation', but with all the partial evaluation
removed. We can mostly use the 'App' type here to get constructors for
the 'IsValue' class operations, except a few of the 'App' constructors
have specialized types (e.g. the 'App' signed extension constructor
'SExt' requires that the sign-extension be non-trivial (i.e. that the
size grows), whereas the 'sext' operation in the 'IsValue' class
allows trivial extensions. In the cases where 'App' is inadequate, we
add constructors to the 'Expr' type instead.

It's not obvious to me that this is better than just reusing the
'IsValue' instance in 'Reopt.Semantics.Implementation' completely: do
we really want to preserve things like '1 + 2', instead of evaluating
them to '3'? Here we preserve all operations.