Implementations of most of evalExpr.
There is a lot of grossness in the floating point helper functions, mostly
because it's difficult to abstract out functions. This is because we use
functions with different types (Float vs Double) in different branches of
a case statement. So, that's gross, but I don't know of a better way.