Only compare up to `Undefined` in `reopt_test`.
This brings the number of non-matching values down to about 22 from
about 57. However, I do mean "about"! The number of non-matches I
observe varies from 21 to 23 now -- it varied from 56 to 58 before
this change -- so we probably think some value in our semantics is
defined when it should be `Undefined`. The sometimes differing value
is always in `%pf` in my limited experiments; I have not looked more
deeply than this:

     diff -u tmp/t21 tmp/t23
    --- tmp/t21	 2015-07-06 17:59:14.750069868 -0700
    +++ tmp/t23	 2015-07-06 18:00:27.577965242 -0700
    @@ -1,3 +1,4 @@
    +pf did not match.  real:  [1]1   emulated: [1]0
     af did not match.  real:  [1]1   emulated: [1]0
     pf did not match.  real:  [1]0   emulated: [1]1
     af did not match.  real:  [1]1   emulated: [1]0
    @@ -18,4 +19,5 @@
     pf did not match.  real:  [1]1   emulated: [1]0
     pf did not match.  real:  [1]0   emulated: [1]1
     pf did not match.  real:  [1]1   emulated: [1]0
    +pf did not match.  real:  [1]0   emulated: [1]1
     af did not match.  real:  [1]0   emulated: [1]1

Here `tmp/t<n>` is generated by

    reopt_test examples/thttpd 2>&1 | grep 'not match'

and contains `<n>` lines.