Implement `evalStmt Syscall`.
By adding a `syscall` operation to `MonadMachineState` and
implementing it in `ConcreteState` by forgetting everything we know
about the machine state. The intuition is that anything could happen
during a syscall.

This gets us up to 563 instructions executed for `examples/thttpd`,
before we die with due to unimplemented `cpuid` instruction:

    II {iiLockPrefix = NoLockPrefix, iiAddrSize = Size64, iiOp = "cpuid", iiArgs = []}
    Unsupported instruction: II {iiLockPrefix = NoLockPrefix, iiAddrSize = Size64, iiOp = "cpuid", iiArgs = []}
    reopt_test: user error (could not exec instruction at [64]4683315)

Looking at the code in `Main_reopt.hs`, it seems that perhaps
`PTraceMachineState` should not be a separate `MonadMachineState`
instance -- it doesn't implement many operations -- but rather should
be merged into `ConcreteState`, or made instead an instance
`MonadMachineStateReader`, a subclass of `MonadMachineState` which
only has `getMem` and `getReg` and `dumpRegs`.