Add support for `xgetbv` instruction.
By treating it as primitive, since its semantics seem to depend on
`cpuid`. When it is supported, its semantics are simple:

    EDX:EAX ← XCR[ECX];

However, I'm not sure we have an extended control register (XCR).

This greatly increases the number of instructions we can handle, and
we now die due to lack of segmentation support:

    $ reopt_test examples/thttpd 2>&1 | grep instruction | wc -l
    1136
    $ reopt_test examples/thttpd 2>&1 | grep "not match" | wc -l
    45

The segmentation error is from
`Reopt.Semantics.FlexdisMatcher.getBVAddress`.