Add support for FS and GS segments in `reopt_test`.
This is very easy because `ptrace` has direct support for FS and GS
base addresses.

This gets us up to 2949 instructions processed, with 76 disagreements
-- I get ~90 disagreements if I add 1 to the number ptrace gives for
`fs_base` and `gs_base`, so it seems this implementation is correct --
and we now fail due to another unsupported instruction, `rdtsc`.

    $ reopt_test examples/thttpd 2>&1 | grep "instruction" | wc -l
    2949
    $ reopt_test examples/thttpd 2>&1 | grep "not match" | wc -l
    76

The `rdtsc` instruction, which returns the "number of cycles since
reset", appears to be another primitive ...