Refactor the division helper from the last commit.
Also, refactor more of the old division code to use the new helper.

I'm leaving the last commit in because it illustrates a GHC bug which
I spent half the day tracking down: GHC fails to deduce a reflexive
type equality `t ~ t` for `t = 4 <=? (n + n)` and other values of
`t`.

Here is a minimal example that triggers the bug:

    {-# LANGUAGE GADTs #-}
    {-# LANGUAGE DataKinds #-}
    {-# LANGUAGE TypeOperators #-}

    module TypeLitsBug where

    import GHC.TypeLits

    data T a where MkT :: T Int

    test :: ((x <=? y) ~ 'True, (y <=? z) ~ 'True)
         => proxy x y z -> ()
    test _ = case MkT of MkT -> ()

The key ingrediants seem to be implicit transitivity of `<=` and GADT
pattern matches.

I reported the bug here: https://ghc.haskell.org/trac/ghc/ticket/10742