Update LLVM translation of division and add documentation.
The LLVM translation remains unchanged, but now there is an
explanation for why it's correct.

Also, I documented but did not fix some bugs -- these bugs existed
before the division refactor. One of them is common -- we use the
wrong bitvector division ops -- and should be fixed. The other -- we
may not properly detect overflow when dividing the largest negative
number by -1 -- is super rare and I'm going to ignore it for now.