Initial support for subgraph comparisons within a function
Currently uses a single register map for all subgraphs - this should be fixable
in the future but we haven't done the engineering work for it.

Some changes to the symbolic semantics to populate location fields that are used
to recognize initial and cut blocks.

Relies on an addition to Crucible.