Port `Reopt.Semantics.Semantics` to new `Location` type.
I eliminated `xmm_low64` and replaced it with explicit "modify the
lower order bits" functions, because `xmm_low64` is not actually
register, unlike e.g. `reg_low32` and `x87reg_mmx`. The idea is that
semantics that are explicit in individual instructions -- described in
Volume 2 of the Intel x86 manual -- should be captured in the
`Semantics` semantics, whereas implicit semantics of aliased registers
(subregisters) -- described in Volume 1 of the Intel x86 manual -- can
be implicit via `Location`. This may be arbitrary ...