Fix buggy implementation of write view for const-upper-bits regs.
I had mixed up the `v` and `v0` arguments. This is rare given all the
`BVType n` indices, and the switch here to uses my generalized `bvCat`
is both simpler and would not type check with the old error.

There are still `reopt_test` mismatches for low8 register writes, so
the default write view is also buggy ...