Move read/write view functions out of `RegisterView`.
Now, instead of embedding the read and write views directly, a
`RegisterView` includes a new `RegisterViewType` tag, specifying what
kind of view it is. External `registerViewRead` and
`registerViewWrite` functions compute the read and write views for
given `RegisterView`. The benefit is that the `RegisterViewType` can
be inspected at runtime, whereas the the read and write views
themselves cannot, since they are functions.

There is some trickery with wrapping the constraints needed to run the
view functions inside the `RegisterViewType` tags, but this is hidden
from the user.

The `RegisterView` type is now an "abstract type", where the
constructor is no longer exposed. Instead, the caller now ses
functions like `registerViewRead` and `registerViewType` to
interrogate a `RegisterView`.