Port `Reopt.Symbolic.Semantics` to the new `Location`.
And add notes about factoring into a new class the basic bitvector
operations needed by the `Location` reads and writes.

From looking at `Reopt.Symbolic.Semantics` a bit here, it seems that
it's very much a work in progress, and will likely break in the future
as we feed more complex examples in `radss_compare`.