Slightly hacky attempt to get more information about syscalls numbers
If the syscall number is an initial register, and that register is
known to be a singleton (via abs int) then we can learn the concrete
syscall no.  The code doesn't transfer values, so any computation on
system call numbers in the block in which the syscall occurs will still
be unknown.