Add `memcpy` + `strlen` example.
This appears to expose a bug in our `libreopt` implementation of
`memcpy`: it appears the copy rounds down to the nearest multiple of 8
bytes. Output from running `llvm-testbed/memcpy-and-strlen/tmp/main`:

    $ tmp/main
    Equalities should hold:
    51 = 51 = 48 = 51 = 51 = 48
    0xc58050 = 0xc58050
    0xc58010 = 0xc58010
    Time flies like the wind; fruit flies like bananas. =
    Time flies like the wind; fruit flies like bananas. =
    Time flies like the wind; fruit flies like banan =
    Time flies like the wind; fruit flies like bananas.
    56 = 56 = 56 = 56 = 56 = 56
    0xc580e0 = 0xc580e0
    0xc58090 = 0xc58090
    Time flies like the wind; fruit flies like bananas.xxxxx =
    Time flies like the wind; fruit flies like bananas.xxxxx =
    Time flies like the wind; fruit flies like bananas.xxxxx =
    Time flies like the wind; fruit flies like bananas.xxxxx

Note that the copy of the 51 byte string in the first example is
truncated to 48 bytes. And now that I'm writing this, did we just get
lucky on the null byte???