Treat `RelocatableValue` as a MemWord instead of a SegmentedAddr
This change was needed because the translation of arithmetic operations on
RelocatableValues (which comes up often in IP-relative addressing) was
generating invalid SegmentedAddrs.  The problem is that the IP lives in the
segment containing the .text section, which is marked as executable.
IP-relative addressing of data adds a large offset to the IP to address a data
item in the non-executable data segment.  Unfortunately, the resulting
SegmentedAddrs thought that their base was in the segment containing .text and
was marking data as executable.  This cause the code discovery pass to try to
disassemble things known to be data as code, which generally failed.

The new translation is just as MemWords, which are given their interpretation as
segmented addresses in the code discovery code (see macaw).  That code has
access to the entire memory, so we can assign the proper segment to them.