Make it easier to run the test suite
This commit checks in the actual binaries that the test suite runs on.  This
means that MacOS users can run the tests without having a cross compilation
toolchain available.  It also keeps the addresses in the binary (i.e., the
binary layout) stable so that the expected results don't need to change
per-system.