Refactoring, and additional error checking.
This migrates some of the name generation code from LLVM rendering to
function recovery, and adds additional checks that must be satisfied
for a function to be recovered.  These changes are intended to help
facilitate VCG generation.

This module also moves the functionality of Reopt.Interface into
Reopt, as the two modules had the same intended purpose of providing a
top level interface to Reopt's capabilities.