vscode: implemented proper persistence of reopt-vcg results
This was harder than it seemed, because there are multiple levels of
persistence.  A webview can have a persisted state, that is saved when the
webview goes out of view.  However, because the webview is not running then, it
cannot update its state when reopt-vcg keeps producing results and the webview
is hidden.

Therefore, the extension itself needs to persist some state and listen to
reopt-vcg while the webview may be inactive.  This is done with the extension's
workspace state.

This now creates another synchronization problem, when the webview is resumed,
as it needs to receive the real authoritative state from the extension since
its persisted state may be stale.

At this point, having these entries in the webview's persisted state may be
pointless.