connect constraint generation and constraint solving
We can now run the pipeline of constraint generation all the way through
constraint solving and display the results of the solving process.

Results are now displaying the generated constraints (which are informative but
in terms of fresh type variables so hard to understand), and the inferred types
for all user-facing variables.

This also solves a bug where we were conflating `FnAssignId`s across
`Function`s, and reengineers the constraint generation monad to be
parameterized by an increasingly-populated context as we go down modules,
functions, and function blocks.