detect the actual tar version present on the system
The current script does not work on my machine because it assumes that any
machine running Darwin will also have bsdtar.  I am running Darwin, but have
also installed the superior GNU tar, so the assumption does not hold.

Instead, I made it read the `--version` output to see if we identify either GNU
or BSD version of tar.