systematically emit a const conversion from int to ptr
For `getelementptr`, we introduced constant expressions turning constant
bitvector values known to be used as pointers into pointer values.

We should do so anytime any constant expression is turned into a pointer.  This
commit makes this more systematic, which results in inlined `inttoptr` calls
for constant bitvectors in other instructions, such as but not limited to the
following:

    %t0 = load i8, i8* inttoptr (i64 4211136 to i8*)

    store i8 1, i8* inttoptr (i64 4211136 to i8*)