don't create two type variables for phi nodes
We were unfortunately creating two separate type variables for phi nodes'
`FnAssignId`s: one where the phi node was emitted, and one the first time the
phi variable was referenced in code.

We now make sure to check for an existing type variable for any `FnAssignId`
before deciding to create a fresh one.