Add --trace-constraint-origins flag
This adds a `--trace-constraint-origins` flag, which can be thought of as a
souped-up version of `--trace-unification` that also prints out the provenance
of each constraint. For now, I have only added the necessary machinery to print
constraints arising from equality constraint (`EqC`). This alone takes quite a
bit of plumbing.  Some of the constraint provenance forms (e.g.,
`FromSubTypeC`) are not terribly descriptive, and if these prove to be not
precise enough, we can always refine them later if need be.