use `InclusiveRange` more, and allow non zero-base segments
In practice, it does not seem too bad to allow non-zero base segments
for residual analysis.  Even though the offsets are not real, they still
make sense for this analysis, so long as multiple offsets aren't reused
in segments that would have different bases, which we could check for if
this becomes an issue.