implement prototype type inference discovery loop
This commit factors out the discovery / type inference code, so that it
may be run repeatedly in a loop until we reach a fixed point.

At each iteration, the type inference may be able to identify certain
addresses as function pointer and figure out their types, which means we
could now recover them and explore more code.

This is not the final version of the algorithm, but a working minimal
version.  More care should be taken in the function type reconstruction.