Use x86CallRegs to resolve function pointers
Instead of reimplementing an incomplete version of x86CallRegs that
works for discovering function pointers in the case of
__libc_start_main, we insert the special case directly into x86CallRegs,
and call that.

We also split x86CallRegs into two parts so that we can retrieve the
name of the function (found by x86CallRegs internally) which is needed
for finding out if a given parameter is a function pointer.