Add case that interprets bitvector as a pointer
When we know that we're dealing with a value that represents an argument
to a function, and that the function has function pointer as a
parameter, we want to add the function pointer's address to Macaw's
exploration frontier.

We noticed that sometimes, Reopt may not even realize that a value it's
dealing with with a pointer at all.
In this case we need to try to interpret it as an absolute address,
and add it to the frontier if the pointer is valid.